Extract graph::tests submodule into own file
